break apart linter module
